Fixed issue of create project needs dockrization

+++ lib/src/cli/commands/create.dart
@@ -6,7 +6,7 @@ import 'package:archive/archive_io.dart';
 
 class CreateProject {
   String projectUrl =
-      'https://github.com/uproid/{template}-finch{docker}/archive/refs/heads/master.zip';
+      'https://github.com/uproid/{template}-finch-docker/archive/refs/heads/master.zip';
   String savePath =
       '${Directory.systemTemp.path}/template_${DateTime.timestamp().microsecondsSinceEpoch}.zip';
 
@@ -42,14 +42,11 @@ class CreateProject {
     if (!Directory(path).existsSync()) {
       return CappConsole("Error creating this path: $path", CappColors.error);
     }
-    var template = controller.getOption('template', def: 'helloworld');
+    var template = controller.getOption('template', def: 'simple');
 
-    var repoUrl = projectUrl
-        .replaceAll('{template}', template)
-        .replaceAll('{docker}', useDocker ? '-docker' : '');
-    var dirName = '{template}-finch{docker}-master'
-        .replaceAll('{template}', template)
-        .replaceAll('{docker}', useDocker ? '-docker' : '');
+    var repoUrl = projectUrl.replaceAll('{template}', template);
+    var dirName =
+        '{template}-finch-docker-master'.replaceAll('{template}', template);
     String pathZip = await CappConsole.progress<String>(
       "Waitng...",
       () async {
@@ -144,13 +141,13 @@ class CreateProject {
 
     if (!useDocker) {
       final dockerFile = File(joinPaths([filePath, 'Dockerfile']));
-      dockerFile.deleteSync();
+      if (dockerFile.existsSync()) dockerFile.deleteSync();
 
       final dockerIgnore = File(joinPaths([filePath, '.dockerignore']));
-      dockerIgnore.deleteSync();
+      if (dockerIgnore.existsSync()) dockerIgnore.deleteSync();
 
       final dockerCompose = File(joinPaths([filePath, 'docker-compose.yaml']));
-      dockerCompose.deleteSync();
+      if (dockerCompose.existsSync()) dockerCompose.deleteSync();
     }
   }
 }
